Crate lastshot

This crate provides an asynchronous "last-shot" channel. A "last-shot" channel is a channel in which the receivers only receive the last message.

This is similar to Tokio's watch. However, Tokio's watch always contain a message, it is not possible to take it. You can still listen for updates, but no matter how many times you read, you won't take the value away. In fact, tokio's watch clones the message if you need it borrowed.

lastshot on the other hand, always take a value away from the channel. And this is always the last sent value. Sending a new value overrides the previous message.

This channel is a multiple producers, multiple consumers channel. Copying Senders and Receivers is the cost of incrementing two atomic integers.

With multiple Receivers, however, some Receivers might get the message in a different order from which they asked for a message.

Perhaps this is not a good name. If you have any ideas, please, suggest it to me.

Structs

NoReceivers
Returned when there are no Receivers connected.

NoSenders
Returned when there are no Senders connected.

Receiver
A receiver handle of a MPMC last-shot channel. It receives messages from the channel. It can be freely cloned and if T: Send, also shared. The cost of cloning is the cost of incrementing two atomic variables.

Sender
A sender handle of a MPMC last-shot channel. It sends messages messages from the channel. It can be freely cloned and if T: Send, also shared. The cost of cloning is the cost of incrementing two atomic variables. Sending a message always overwrite the previous message.

Functions

channel
Creates a new multiple producers multiple consumers (MPMC) "last-shot" channel.